pyproject.toml
[tool.poetry]
name = "fastapi-gsheets"
version = "0.1.0"
description = ""
authors = ["modernpacifist <modernpacifist@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
fastapi = "^0.111.1"
google-api-python-client = "^2.137.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
